extract battle hook
