+++ modules/ui/src/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from "./positionStyles";
 export * from "./useDemo";
+export * from "./useBattle";
